requirements.txt
# Streamlit UI
streamlit==1.35.0

# AWS SDK
boto3==1.34.145
botocore==1.34.145

# OpenAI SDK (latest >=1.0.0)
openai==1.25.0

# Environment variables
python-dotenv==1.0.1

# JSON validation/manipulation
jsonschema==4.22.0

# Other utilities
requests==2.31.0
